+++ app/models/subtask.rb
@@ -6,7 +6,7 @@ class Subtask < ActiveRecord::Base
   belongs_to  :user
   belongs_to  :task
 
-  validates :title, presence: true, length: { minimum: 1 }
+  validates :title, presence: true, length: { minimum: 3 }
   validates :user_id, presence: true
   validates :task_id, presence: true
