+++ spec/models/subtask_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
 describe Subtask do
-  pending "add some examples to (or delete) #{__FILE__}"
+
 end
